+++ src/REF_MMult.c
@@ -6,7 +6,7 @@
 
 /* Routine for computing C = A * B + C */
 
-void REF_MMult( int m, int n, int k, float *a, int lda, float *b, int ldb, float *c, int ldc ){
+void REF_MMult( int m, int n, int k, double *a, int lda, double *b, int ldb, double *c, int ldc ){
     int i, j, p;
     for ( i=0; i<m; i++ ){
         for ( j=0; j<n; j++ ){
