+++ src/copy_matrix.c
@@ -1,7 +1,7 @@
 #define A( i, j ) a[ (j)*lda + (i) ]
 #define B( i, j ) b[ (j)*ldb + (i) ]
 
-void copy_matrix( int m, int n, float *a, int lda, float *b, int ldb )
+void copy_matrix( int m, int n, double *a, int lda, double *b, int ldb )
 {
   int i, j;
